+++ src/AngleSharp/Css/Parser/CssCombinator.cs
@@ -24,9 +24,9 @@ namespace AngleSharp.Css.Parser
         public static readonly CssCombinator Deep = new DeepCombinator();
 
         /// <summary>
-        /// The descendent operator (space, or alternatively >>).
+        /// The descendant operator (space, or alternatively >>).
         /// </summary>
-        public static readonly CssCombinator Descendent = new DescendentCombinator();
+        public static readonly CssCombinator Descendant = new DescendantCombinator();
 
         /// <summary>
         /// The adjacent sibling combinator +.
@@ -115,11 +115,11 @@ namespace AngleSharp.Css.Parser
             }
         }
 
-        private sealed class DescendentCombinator : CssCombinator
+        private sealed class DescendantCombinator : CssCombinator
         {
-            public DescendentCombinator()
+            public DescendantCombinator()
             {
-                Delimiter = CombinatorSymbols.Descendent;
+                Delimiter = CombinatorSymbols.Descendant;
                 Transform = el =>
                 {
                     var parents = new List<IElement>();