+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+# 1.0.5
+
+ Released on Tuesday, October 3 2023
+
+ - Improved codebase (#1128, #1126, #1133) @matkoch @ivandrofly
+ - Improved documentation (#1127)
+ - Improved child selector performance (#1135) @lahma
+ - Improved query selector all performance (#584, #1134) @lahma
+ - Fixed issue with foreign content using end tags
+ - Fixed typo `descendants` (#1136) @SimonCropp
+ - Added a way to resolve symbols by their name to `HtmlEntityProvider` (#396)
+
 # 1.0.4
 
 Released on Saturday, June 24 2023